M15: one CGRect — alias the geometry types to Foundation's
CGFloat/CGPoint/CGSize/CGVector/CGRect are now typealiases to Foundation's
types on any platform that has Foundation, so a program that imports both
OpenUIKit and Foundation sees exactly ONE of each instead of two colliding
names. CGAffineTransform stays ours (Foundation has none), and the two members
Foundation lacks — CGPoint.applying / CGRect.applying — are re-added as
extensions with the original semantics.

Darwin needs a plain `import CoreGraphics` in Geometry.swift: Foundation
re-exports the CG structs but the Swift members (minX, insetBy, integral, ...)
live in the CoreGraphics overlay. It is deliberately not `@_exported`, so
CoreGraphics' own CGColor/CGLineCap/CGAffineTransform do not leak in and
collide with ours.

Fallout, all of it the CGFloat-is-no-longer-Double hazard:
  * Rasterizer's `extension CGFloat { clamped }` silently covered Double;
    it is now `extension FloatingPoint`.
  * Six sites needed an explicit CGFloat where inference used to be free.

108/108 scenes, 9/9 traces, 732 tests green.

+++ Sources/OpenUIKit/UIStackView.swift
@@ -265,7 +265,7 @@ open class UIStackView: UIView {
             let flexible = visiblePairs.indices.filter { isAxisFlexible(visiblePairs[$0].0) }
             if let lastFlexible = flexible.last {
                 for i in flexible { lengths[i] = 0 }
-                let slack = available - lengths.reduce(0, +)
+                let slack = available - lengths.reduce(CGFloat(0), +)
                 if slack >= 0 {
                     lengths[lastFlexible] = slack
                 } else if let firstRigid = lengths.indices.first(where: { !flexible.contains($0) }) {
